requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
